rename farm controller file name
